knapsack_heuristic_calculator_analyzer/requirements.txt
numpy>=1.20.0
pandas>=1.3.0
numba>=0.54.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pytz>=2021.1
psutil>=5.8.0
py-cpuinfo>=8.0.0
